First version of Dynamic Pull Timer
AceConsole is currently not needed as there is no slash cmd, but
I kept it because it might be needed in future versions.

The first version creates the dynamic pull timer as a minimap icon
on which a left click starts the pull timer, a right click opens
the configruation window.